requirements.txt
fastapi==0.103.1
loguru==0.7.0
uvicorn==0.23.2
uvloop==0.17.0
python-dotenv==1.0.0